Cambios en el modelo

+++ AppCityParkServices/AppCityParkServices/Models/SalarioBasico.cs
@@ -22,7 +22,7 @@ namespace AppCityParkServices.Models
     
         public int SalarioBasicoId { get; set; }
         public string Descripcion { get; set; }
-        public Nullable<double> Monto { get; set; }
+        public Nullable<decimal> Monto { get; set; }
         public System.DateTime Fecha { get; set; }
         public Nullable<int> EmpresaId { get; set; }
